refactor(core): move SerializablePkg to serde_v8
Also rename `MinValue` to `Primitive`
